# Add bit packing to NodeImpl