@chandlerc 2024-12-16
Yeah, let me work on getting some benchmark data, both for this PR and the other PR on x86. I got distracted by something else on Friday, but should be able to get you good data today. And as you point out, good to know whether what I found on Arm64 reproduces on x86...