@chandlerc 2024-12-16
A bit closer to the noise floor than I would like, but does seem like a consistent win on x86:

```
BM_CompileAPIFileDenseDecls<Phase::Parse>/256       189k ± 1%            187k ± 1%  -1.23%  (p=0.000 n=19+20)
BM_CompileAPIFileDenseDecls<Phase::Parse>/1024      893k ± 1%            878k ± 2%  -1.64%  (p=0.000 n=20+20)
BM_CompileAPIFileDenseDecls<Phase::Parse>/4096     3.63M ± 1%           3.55M ± 1%  -2.23%  (p=0.000 n=16+19)
BM_CompileAPIFileDenseDecls<Phase::Parse>/16384    15.3M ± 2%           15.0M ± 0%  -1.91%  (p=0.000 n=20+16)
BM_CompileAPIFileDenseDecls<Phase::Parse>/65536    63.6M ± 1%           62.6M ± 0%  -1.63%  (p=0.000 n=20+16)
BM_CompileAPIFileDenseDecls<Phase::Parse>/262144    276M ± 1%            273M ± 1%  -1.21%  (p=0.000 n=20+18)
```

On a Neoverse-N1-ish Arm server, bit less clear but still seems positive:

```
BM_CompileAPIFileDenseDecls<Phase::Parse>/256        180k ± 1%            180k ± 0%    ~     (p=0.076 n=20+20)
BM_CompileAPIFileDenseDecls<Phase::Parse>/1024       862k ± 0%            861k ± 0%  -0.13%  (p=0.001 n=19+18)
BM_CompileAPIFileDenseDecls<Phase::Parse>/4096      3.55M ± 0%           3.54M ± 0%  -0.32%  (p=0.000 n=20+20)
BM_CompileAPIFileDenseDecls<Phase::Parse>/16384     15.1M ± 0%           15.0M ± 0%  -0.16%  (p=0.006 n=20+20)
BM_CompileAPIFileDenseDecls<Phase::Parse>/65536     65.6M ± 0%           65.5M ± 1%  -0.25%  (p=0.000 n=16+20)
BM_CompileAPIFileDenseDecls<Phase::Parse>/262144     300M ± 1%            298M ± 1%  -0.87%  (p=0.000 n=20+20)
```